Moved common logic into base excel class https://github.com/SoftUni-Internal/exam-systems-issues/issues/366
