+++ Open Judge System/Services/OJS.Services/OJS.Services.csproj
@@ -110,9 +110,10 @@
     <Compile Include="Business\ExamGroups\IExamGroupsBusinessService.cs" />
     <Compile Include="Business\ParticipantScores\IParticipantScoresBusinessService.cs" />
     <Compile Include="Business\ParticipantScores\Models\CategoryContestsParticipationSummary.cs" />
-    <Compile Include="Business\ParticipantScores\Models\CategoryContestsParticipationSummaryExcel.cs" />
-    <Compile Include="Business\ParticipantScores\Models\ContestParticipationSummaryExcel.cs" />
     <Compile Include="Business\ParticipantScores\Models\ContestParticipationSummary.cs" />
+    <Compile Include="Business\ParticipantScores\Models\ExcelModels\CategoryContestsParticipationSummaryExcel.cs" />
+    <Compile Include="Business\ParticipantScores\Models\ExcelModels\ContestParticipationSummaryExcelModel.cs" />
+    <Compile Include="Business\ParticipantScores\Models\ExcelModels\BaseParticipationStatisticsSummaryExcelModel.cs" />
     <Compile Include="Business\ParticipantScores\Models\MaximumResultSubmissionByProblemServiceModel.cs" />
     <Compile Include="Business\ParticipantScores\Models\ParticipantScoresSummaryModel.cs" />
     <Compile Include="Business\ParticipantScores\Models\ParticipantSummaryInfoServiceModel.cs" />
