+++ Open Judge System/Web/OJS.Web/Properties/PublishProfiles/Dev Suls Judge System.pubxml
@@ -8,11 +8,11 @@ by editing this MSBuild file. In order to learn more about this please visit htt
     <WebPublishMethod>MSDeploy</WebPublishMethod>
     <LastUsedBuildConfiguration>Dev</LastUsedBuildConfiguration>
     <LastUsedPlatform>Any CPU</LastUsedPlatform>
-    <SiteUrlToLaunchAfterPublish>https://dev.platform.softuni.bg:92/</SiteUrlToLaunchAfterPublish>
+    <SiteUrlToLaunchAfterPublish>https://dev.judge.softuni.org/</SiteUrlToLaunchAfterPublish>
     <LaunchSiteAfterPublish>True</LaunchSiteAfterPublish>
     <ExcludeApp_Data>False</ExcludeApp_Data>
-    <MSDeployServiceURL>217.174.159.198</MSDeployServiceURL>
-    <DeployIisAppPath>dev.judge.softuni.bg</DeployIisAppPath>
+    <MSDeployServiceURL>217.174.159.130</MSDeployServiceURL>
+    <DeployIisAppPath>dev.judge.softuni.org</DeployIisAppPath>
     <RemoteSitePhysicalPath />
     <SkipExtraFilesOnServer>True</SkipExtraFilesOnServer>
     <MSDeployPublishMethod>WMSVC</MSDeployPublishMethod>
